Extract AST node types to separate file
