+++ src/core/Formatter.ts
@@ -3,10 +3,11 @@ import Params from './Params';
 import Tokenizer from './Tokenizer';
 import formatCommaPositions from './formatCommaPositions';
 import formatAliasPositions from './formatAliasPositions';
-import Parser, { type Statement } from './Parser';
+import Parser from './Parser';
 import ExpressionFormatter from './ExpressionFormatter';
 import { indentString } from './config';
 import AliasAs from './AliasAs';
+import { Statement } from './ast';
 
 /** Main formatter class that produces a final output string from list of tokens */
 export default class Formatter {
